// frontend/src/openapi/models/CreateFeatureStrategySchema.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Unleash API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 4.15.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { ConstraintSchema } from './ConstraintSchema';
import {
    ConstraintSchemaFromJSON,
    ConstraintSchemaFromJSONTyped,
    ConstraintSchemaToJSON,
} from './ConstraintSchema';

/**
 * 
 * @export
 * @interface CreateFeatureStrategySchema
 */
export interface CreateFeatureStrategySchema {
    /**
     * 
     * @type {string}
     * @memberof CreateFeatureStrategySchema
     */
    name: string;
    /**
     * 
     * @type {number}
     * @memberof CreateFeatureStrategySchema
     */
    sortOrder?: number;
    /**
     * 
     * @type {Array<ConstraintSchema>}
     * @memberof CreateFeatureStrategySchema
     */
    constraints?: Array<ConstraintSchema>;
    /**
     * 
     * @type {{ [key: string]: string; }}
     * @memberof CreateFeatureStrategySchema
     */
    parameters?: { [key: string]: string; };
}

/**
 * Check if a given object implements the CreateFeatureStrategySchema interface.
 */
export function instanceOfCreateFeatureStrategySchema(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "name" in value;

    return isInstance;
}

export function CreateFeatureStrategySchemaFromJSON(json: any): CreateFeatureStrategySchema {
    return CreateFeatureStrategySchemaFromJSONTyped(json, false);
}

export function CreateFeatureStrategySchemaFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateFeatureStrategySchema {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'name': json['name'],
        'sortOrder': !exists(json, 'sortOrder') ? undefined : json['sortOrder'],
        'constraints': !exists(json, 'constraints') ? undefined : ((json['constraints'] as Array<any>).map(ConstraintSchemaFromJSON)),
        'parameters': !exists(json, 'parameters') ? undefined : json['parameters'],
    };
}

export function CreateFeatureStrategySchemaToJSON(value?: CreateFeatureStrategySchema | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'name': value.name,
        'sortOrder': value.sortOrder,
        'constraints': value.constraints === undefined ? undefined : ((value.constraints as Array<any>).map(ConstraintSchemaToJSON)),
        'parameters': value.parameters,
    };
}